requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
graphviz==0.21
pydotplus==2.0.2
PyJWT==2.12.1
pyparsing==3.3.2
sqlparse==0.5.5
